XRisk {XRisk}	R Documentation

XRisk: Tools for risk assessments using Xplicit simulation data

Description

XRisk provides means of accessing data stored by Xplicit data providers in R. Currently, the HDF5-based x3df format is supported and data loaded from there can be aggregated over scales and in chunks, allowing large volumes to be processed in R. XRisk does also allow to define statistical populations and apply functions by these statistical populations on data in R data.tables. Finally, XRisk contains helper functions to read and write spatial and tabular data.

Data access

Access to Xplicit simulation data is organized by R6 classes:

X3DF

Represents a X3DF data source and provides methods to explore its contents.

X3dfDataset

A individual dataset from a X3DF data source. Includes methods to retrieve data and metadata from the dataset.

Statistical populations

Statistical populations are prepared by a set of functions:

statPop

Compiles a data.table containing definitions for statistical populations.

allLevels

Generates criteria to consider every unique level of a variable in a data.table as statistical population.

intervals

Generates criteria to consider numerical interval ranges of a variable as statistical populations.

spapply

Applies a function to all sttistical populations within a data.table.

Input and output

XRisk provides functions helping with input and output of data:

readRasterMap

Reads data from a raster file and outputs it in a data.table.

writeRasterMap

Writes data to a raster file.

writeToXlsx

Writes data to an xlsx workbook.

[Package XRisk version 0.0.0.9002 Index]