allLevels {XRisk}	R Documentation

Filter criteria for all levels of a variable

Description

allLevles generates a named character vector containing one filter criteria for every level of a variable that can be used in the statPop function.

Usage

allLevels(x, col, add.na = NULL)

Arguments

x
A data.table.

col
A unevaluated expression. The name of a column of x.

add.na
A scalar character. If not null, adds a criterium that selects all elements. Defaults to NULL.

Details

col is evaluated in the context of x and reduced to its unique values. For each unique value, a character string in the format "(a == b)" is concatenated where a is the substitue of col and b is the unique value. The names of the resulting vector are set to the unique values. If a add.na is provided, it is prepended to the result as the value TRUE with name add.na.

Value

A named vector containing individual filter criteria for usage in the statPop function. The names of the vector equal the individual levels of x.

Examples

x <- data.table::data.table(group = sample(1:5, 100, TRUE))
allLevels(x, group, "all")

[Package XRisk version 0.0.0.9002 Index]