intervals {XRisk}	R Documentation

Filter criteria for intervals

Description

intervals generates a named character vector containing filter criteria for predefined intervals that can be used in the statPop function.

Usage

intervals(col, lower, upper, add.na = NULL)

Arguments

col
A character scalar. The name of the varaiable to which the criterium applies.

lower
A numeric vector. The lower, exclusive boundaries of the intervals.

upper
A numeric vector of the same length as lower. The upper, inclusive boundaries of the intervals.

add.na
A scalar character. If not null, adds a criterium that selects all elements. Defaults to NULL.

Details

Criteria are outputted in the format "(a > b & c <= b)" where a are the values of lower, b the values of upper and c equals col. The names of the resulting vector are prepared using the format "lower-upper" where lower and upper are replaced accordingly. If a add.na is provided, it is prepended to the result as the value TRUE with name add.na.

Value

A named vector containing individual filter criteria for usage in the statPop function.

Examples

intervals("len", c(1, 1.5, 2), c(1.5, 2, 2.5), "any")

[Package XRisk version 0.0.0.9002 Index]