readRasterMap {XRisk}	R Documentation

Read raster map

Description

readRasterMap loads the values of a raster along with its according x- and y-coordinates into a data.table.

Usage

readRasterMap(x, val = "value")

Arguments

x
A character scalar. The path of the raster file.

val
A character scalar. The name of the value column. Defaults to "value".

Details

The raster package must be installed for this function to work. It is then used to read the values of the raster into a data.table. x- and y-coordinates are bound to the resulting data.table as additional columns. Coordinates are given as column and row indices of the raster.

Value

Returns a data.table with three columns: x-coordinates as raster column indices, y-coordinates as raster row indices and values as loaded from the raster.

Examples

## Not run: 
readRasterMap("C:Temp/LULC/LulcRaster.tif", "lulc")

## End(Not run)

[Package XRisk version 0.0.0.9002 Index]