spapply {XRisk}	R Documentation

Applying to statistical populations

Description

spapply applies a function to statistical populations retrieved from a data.table.

Usage

spapply(x, statPop, fun)

Arguments

x
A data.table. Contains the entire data that is analyzed.

statPop
A data.table. Specifies individual statistical populations to which fun is applied. Commonly the output of the statPop function.

fun
A function. The function that is applied to each statistical population defined by statPop.

Details

fun is called for every statistical population in x. The statistical population is retrieved from x by evaluating the filter criteria stored in a column .def within statPop. The result value of the function is prepended by the other columns present in statPop to identify the statistical population the result value applies to.

Value

A list of data.tables. Each list item contains the result for one statistical population.

Examples

data <- data.table::data.table(
  group = sample(1:5, 100, TRUE),
  len = rnorm(100, 1.5, .5),
  val = rnorm(100)
)
sp <- statPop(
  group = allLevels(data, group, "all"),
  len = intervals("len", c(1, 1.5, 2), c(1.5, 2, 2.5), "any")
)
data.table::rbindlist(
  spapply(data, sp, function(dt) { dt[, .(mean = mean(val))] }))

[Package XRisk version 0.0.0.9002 Index]