statPop {XRisk}	R Documentation

Define statistical populations

Description

statPop defines statistical populations for use in the spapply function.

Usage

statPop(...)

Arguments

...
Two or more named character vectors. Each vector contains an individual filter condition that is included in the definition of a statistical population.

Details

Statistical populations are understood here as subsets of data, e.g. filtered rows of a data.table. In xrisk, analyses are often performed for many, partly overlapping statistical populations. The statPop function prepares these statistical populations by defining filter criterias for the cartesian product of individual conditions.

Value

A data.table with a .def column that contains the unevaluated filter criterium for the statistical population. It consists of the individual criteria concatenated bv "&". For each individual criterium, a further column is present in the returned data.table that has the name of the argument it was passed as to the function. These columns contains the names of the indiviudal criteria that are considered in the regarding statistical population.

Examples

a <- paste("a ==", 1:2)
names(a) <- paste("group", 1:2)

b <- paste("b >", 2:4)
names(b) <- b

statPop(A = a, B = b)

[Package XRisk version 0.0.0.9002 Index]