writeRasterMap {XRisk}	R Documentation

Write raster map

Description

writeRasterMap writes a vector to a raster.

Usage

writeRasterMap(x, output, extent, crs, resolution = 1)

Arguments

x
A numerical vector. Contains the data written to the raster.

output
A character scalar. The file path to the non-existing file to which the data will be written.

extent
A numerical vector of length 4. The extent of the raster given by coordinates in the specified crs. Reads as xmin, xmax, ymin, ymax.

crs
A character scalar. Either a Proj4 or CRS representation of the spatial coordinate system of the output raster.

resolution
A numeric scalar. The resolution of the raster in crs units. Defaults to 1.

Details

The raster package must be installed for this function to work. Reference systems may be given either in Proj4 or WKT for. Internally, CRS given in WKT will be transformed to Proj4 and the rgdal package must be installed to perform this transformation. The number of values of x must equal the number of cells in the raster. Latter is determined by the given extent and resolution.

Examples

## Not run: 
extent <- c(353865, 353875, 5660637, 5660647)
crs <- "+proj=utm +zone=32 +ellps=GRS80 +units=m +no_defs"
writeRasterMap(runif(100), "C:/Temp/output.tif", extent, crs)

## End(Not run)

[Package XRisk version 0.0.0.9002 Index]