writeToXlsx {XRisk}	R Documentation

Write xlsx file

Description

writeToXlsx writes data into xlsx workbooks.

Usage

writeToXlsx(data, file, author = "")

Arguments

data
A list. Contains the data to write to the xlsx workbook.

file
A character scalar. The file path to the output file.

author
A character scalar. The author that is added as a metadatum to the workbook.

Details

The openxlsx package must be installed for this function to work. The function writes different kind of data into a xlsx workbook. Every item of x is added as an own worksheet to the workbook where the names of x will be used as the names of the worksheets. If the data to be added to a worksheet is a character vector, each element of the vector will be written as an own paragraph into a cell in the first column. Paragraphs are separated by blank rows. Text is wrapped to 80 characters and each line will be written to a different row. Data given as a data.frame is passed as a whole to the writeDataTable function. Other types of data are written elementwise into the worksheet. The author is added as a metadatum to the workbook.

Examples

## Not run: 
writeToXlsx(
  list(info = c("Test", "This is sample output"), data = runif(100)),
  "C:/Temp/sample.xlsx",
  "Author Sample"
)

## End(Not run)

[Package XRisk version 0.0.0.9002 Index]