X3DF {XRisk}	R Documentation

X3DF data source

Description

An X3DF R6Class represents an indivudal X3DF data source.

Usage

X3DF

Format

A R6Class generator object

Details

X3DF classes provide wrappers to individual X3DF data sources given their file system location. The datasets contained by the data source can be listed and accessed.

Fields

datasets

A list of all datasets provided by the data source. Each dataset is provided as a X3dfDataset.

Methods

new(path)

Creates a new X3DF class. The path specifies the file location of the X3DF dataset.

close()

Closes the file handle to the X3DF data source and should be called when the work with the data source is finished.

[Package XRisk version 0.0.0.9002 Index]