X3dfDataset {XRisk}	R Documentation

X3DF dataset

Description

An X3dfDataset R6Class represents an indivudal X3DF dataset from a X3DF data source.

Usage

X3dfDataset

Format

A R6Class generator object

Details

X3dfDataset classes provide wrappers around X3DF datasets. They offer methods to examine the metadata of the dataset and to read the data of the dataset either completely or in chunks while applying a function to the individual chunks.

Fields

name

The name of the dataset.

h5type

The native H5 type with which the dataset is stored.

size

The size of the dataset per scale.

chunksize

The size of chunks with which the data was stored.

x3type

The X3DF data type that was used to store the data.

scales

The scales to which the data refer.

max

The maximum value found in the dataset.

data

The data itself.

Methods

new(d)

Creates a new X3dfDataset class around a H5 dataset reference d. X3dfDataset objects are usually not created manually but by X3DF objects so that H5 handles do not have to be managed by the user.

print()

Produces a formatted output of the object that shows the metadata of the X3DF data set.

chunks(size)

Creates a chunk table with the individual chunks to be considered. size is a numeric vector that specifies the chunk size for every dimension of the dataset. If an element of the vector is specified as NA, the chunksize used when storing the data is used for the according dimension. A size of Inf results in chunks that span all elements of the according dimension.

chapply(over, fun, ..., chunks = NULL, level = "level")

Applies the function fun over specified dimensions to every chunk of the X3DF dataset. over is a character vector that specifies over which dimensions the function is applied. Its elements must be some of the dataset's scales. The ... arguments are passed to the function fun. chunks specifies the sizes of individual chunks and is passed to the object's chunks method to retrieve a chunk table. If chunks is NULL, Inf will be passed for every dimension in over and NA for every other dimension. If fun returns a named vector, the result of chapply will contain a column with the name of level that contains the names of the function's result vector. chapply returns a data.table with the coordinates of the scales that were not included in over, a value column with the function's result and, possibly, a level column.

extract()

Extracts the value for the coordinates given in the coords vector from the dataset. If a single element of the vector is NA, all vectors along the according axis are returned. The return value is a data.table containing coordinates and values.

[Package XRisk version 0.0.0.9002 Index]